+++ src/nodes/cell/SFXCell.hpp
@@ -3,7 +3,7 @@
 #include <memory>
 #include <Geode/Geode.hpp>
 #include <Geode/utils/web.hpp>
-#include "../Utils.hpp"
+#include "../../Utils.hpp"
 
 using namespace geode::prelude;